+++ packages/nimble-components/src/table-column/number-text/models/table-column-number-text-validator.ts
@@ -2,7 +2,11 @@ import type { ColumnInternals } from '../../base/models/column-internals';
 import { ColumnValidator } from '../../base/models/column-validator';
 import { NumberTextFormat } from '../types';
 
-const numberTextValidityFlagNames = ['invalidDecimalDigits'] as const;
+const numberTextValidityFlagNames = [
+    'invalidDecimalDigits',
+    'invalidDecimalMaximumDigits',
+    'decimalDigitsMutuallyExclusiveWithDecimalMaximumDigits'
+] as const;
 
 // The maximum and minimum allowed configuration for 'maximumFractionDigits'
 // and 'minimumFractionDigits' on the NumberFormat.
@@ -29,6 +33,34 @@ export class TableColumnNumberTextValidator extends ColumnValidator<
             ? this.isInvalidDecimalDigitsValue(decimalDigits)
             : false;
         this.setConditionValue('invalidDecimalDigits', invalid);
+    }
+
+    public validateDecimalMaximumDigits(
+        format: NumberTextFormat,
+        decimalMaximumDigits: number | undefined
+    ): void {
+        const shouldValidateDecimalDigitsValue = format === NumberTextFormat.decimal
+            && typeof decimalMaximumDigits === 'number';
+        const invalid = shouldValidateDecimalDigitsValue
+            ? this.isInvalidDecimalDigitsValue(decimalMaximumDigits)
+            : false;
+        this.setConditionValue('invalidDecimalMaximumDigits', invalid);
+    }
+
+    public validateNoMutuallyExclusiveProperties(
+        format: NumberTextFormat,
+        decimalDigits: number | undefined,
+        decimalMaximumDigits: number | undefined
+    ): void {
+        const shouldValidateMutuallyExclusiveProperties = format === NumberTextFormat.decimal;
+        const invalid = shouldValidateMutuallyExclusiveProperties
+            ? typeof decimalDigits === 'number'
+              && typeof decimalMaximumDigits === 'number'
+            : false;
+        this.setConditionValue(
+            'decimalDigitsMutuallyExclusiveWithDecimalMaximumDigits',
+            invalid
+        );
     }
 
     private isInvalidDecimalDigitsValue(decimalDigits: number): boolean {